CRUD

+++ src/main/java/com/example/demo2/entity/Bloc.java
@@ -17,7 +17,7 @@ import java.util.Set;
 public class Bloc implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idBloc;
+    private long idBloc;
     private String nomBloc;
     private long capaciteBloc;
     @ManyToOne
