+++ src/main/java/com/example/demo2/entity/Universite.java
@@ -16,7 +16,7 @@ import java.io.Serializable;
 public class Universite implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idUniversite;
+    private long idUniversite;
     private String nomUniversite;
     private String adresse;
     @OneToOne
